Remove old render engine module
